Code refactoring, resolving compile problems with SASS imports.
