+++ src/app/theming/theme-chooser/theme-chooser-sample.component.scss
@@ -1,4 +1,4 @@
-@import 'themes';
+@import 'styles/theme-classes';
 
 .addProdBtn {
     margin-bottom: 5px;
